fix: menu{}: now namespaced
- Made Menu class a namespace.
- Moved drawAscii() into a new area.
- Various clang-format cleanups.
