+++ src/headers/global.hpp
@@ -7,6 +7,7 @@ using ull = unsigned long long;
 void getInput(char &);
 void newline(int n = 1);
 void clearScreen();
+void drawAscii();
 std::string secondsFormat(double);
 
 #endif
