+++ src/headers/menu.hpp
@@ -1,18 +1,8 @@
 #ifndef MENU_H
 #define MENU_H
 
-class Menu {
-private:
-  void startGame();
-  void showScores();
-  void printMenu();
-  void input(int err);
-  void continueGame();
-
-public:
-  void startMenu(int err = 0);
-};
-
-void drawAscii();
+namespace Menu {
+void startMenu(int err = 0);
+} // namespace Menu
 
 #endif
