+++ src/menu.cpp
@@ -1,37 +1,31 @@
 #include "menu.hpp"
 #include "color.hpp"
 #include "game.hpp"
+#include "global.hpp"
 #include "scores.hpp"
 #include <iostream>
 #include <sstream>
 
-void Menu::startMenu(int err) {
-  constexpr auto greetings_text = "Welcome to ";
-  constexpr auto gamename_text = "2048!";
-  constexpr auto sp = "  ";
+namespace {
 
-  std::ostringstream str_os;
-  std::ostringstream title_richtext;
-  title_richtext << bold_on << sp << greetings_text << blue << gamename_text
-                 << def << bold_off << "\n";
-
-  constexpr auto menu_entry_text = R"(
-          1. Play a New Game
-          2. Continue Previous Game
-          3. View Highscores and Statistics
-          4. Exit
-
-)";
-
-  clearScreen();
-  drawAscii();
-  str_os << title_richtext.str();
-  str_os << menu_entry_text;
-  std::cout << str_os.str();
-  input(err);
+void startGame() {
+  Game g;
+  g.startGame();
 }
 
-void Menu::input(int err) {
+void continueGame() {
+  Game g;
+  g.continueGame();
+}
+
+void showScores() {
+  Scoreboard s;
+  s.printScore();
+  s.printStats();
+}
+
+void input(int err) {
+  using Menu::startMenu;
   constexpr auto err_input_text = "Invalid input. Please try again.";
   constexpr auto prompt_choice_text = "Enter Choice: ";
   constexpr auto sp = "  ";
@@ -74,38 +68,34 @@ void Menu::input(int err) {
   }
 }
 
-void Menu::startGame() {
+} // namespace
 
-  Game g;
-  g.startGame();
+namespace Menu {
+
+void startMenu(int err) {
+  constexpr auto greetings_text = "Welcome to ";
+  constexpr auto gamename_text = "2048!";
+  constexpr auto sp = "  ";
+
+  std::ostringstream str_os;
+  std::ostringstream title_richtext;
+  title_richtext << bold_on << sp << greetings_text << blue << gamename_text
+                 << def << bold_off << "\n";
+
+  constexpr auto menu_entry_text = R"(
+          1. Play a New Game
+          2. Continue Previous Game
+          3. View Highscores and Statistics
+          4. Exit
+
+)";
+
+  clearScreen();
+  drawAscii();
+  str_os << title_richtext.str();
+  str_os << menu_entry_text;
+  std::cout << str_os.str();
+  input(err);
 }
 
-void Menu::continueGame() {
-  Game g;
-  g.continueGame();
-}
-
-void Menu::showScores() {
-
-  Scoreboard s;
-  s.printScore();
-  s.printStats();
-}
-
-void drawAscii() {
-  constexpr auto title_card_2048 = R"(
-   /\\\\\\\\\          /\\\\\\\                /\\\         /\\\\\\\\\
-  /\\\///////\\\      /\\\/////\\\            /\\\\\       /\\\///////\\\
-  \///      \//\\\    /\\\    \//\\\         /\\\/\\\      \/\\\     \/\\\
-             /\\\/    \/\\\     \/\\\       /\\\/\/\\\      \///\\\\\\\\\/
-           /\\\//      \/\\\     \/\\\     /\\\/  \/\\\       /\\\///////\\\
-         /\\\//         \/\\\     \/\\\   /\\\\\\\\\\\\\\\\   /\\\      \//\\\
-        /\\\/            \//\\\    /\\\   \///////////\\\//   \//\\\      /\\\
-        /\\\\\\\\\\\\\\\   \///\\\\\\\/              \/\\\      \///\\\\\\\\\/
-        \///////////////      \///////                \///         \/////////
-  )";
-  std::ostringstream title_card_richtext;
-  title_card_richtext << green << bold_on << title_card_2048 << bold_off << def;
-  title_card_richtext << "\n\n\n";
-  std::cout << title_card_richtext.str();
-}
+} // namespace Menu
